+++ tests/support/editorHarness.ts
@@ -1,4 +1,4 @@
 export { createAnkiCardEditorExtension } from '../../src/editorExtension';
-export { Component, editorInfoField, editorLivePreviewField, setLivePreview, TFile, Menu } from './obsidianMock';
+export { Component, editorInfoField, editorLivePreviewField, setLivePreview, TFile, Menu, MarkdownRenderer } from './obsidianMock';
 export { AnkiCardAutoCompleter } from '../../src/autocomplete';
 export { updateCard, toggleCardRegistration, changeCardType } from '../../src/cardActions';